+++ packages/db/src/schema/index.ts
@@ -23,3 +23,4 @@ export * from './email-suppressions';
 export * from './email-messages';
 export * from './user-totp';
 export * from './flashcards';
+export * from './user-snippets';